Move helper functions to separate utils.rs file
