+++ src/commands.rs
@@ -2,81 +2,16 @@ use crate::config::{
     get_embedded_colorscheme_dir, get_embedded_hooks_dir, BASE16_SHELL_CONFIG_PATH_ENV,
     BASE16_SHELL_THEME_NAME_PATH_ENV, BASE16_THEME_ENV,
 };
+use crate::utils::{read_file_to_string, write_to_file};
 use anyhow::{Context, Result};
 use include_dir::Dir;
-use std::fs::{self, File};
-use std::io::Read;
+use std::fs;
 use std::io::Write;
 use std::os::unix::fs as unix_fs;
 use std::path::{Path, PathBuf};
 use std::process::Command;
 use std::str::from_utf8;
 use tempfile::NamedTempFile;
-
-/// Ensures that a directory exists, creating it if it does not.
-fn ensure_directory_exists<P: AsRef<Path>>(dir_path: P) -> Result<()> {
-    let path = dir_path.as_ref();
-
-    if !path.exists() {
-        fs::create_dir_all(path)
-            .with_context(|| format!("Failed to create directory at {:?}", path))?;
-    }
-
-    Ok(())
-}
-
-/// Ensures that a file exists, creating it if it does not.
-fn ensure_file_exists<P: AsRef<Path>>(file_path: P) -> Result<()> {
-    let path = file_path.as_ref();
-
-    if !path.exists() {
-        File::create(path).with_context(|| format!("Failed to create file at {:?}", path))?;
-    };
-
-    Ok(())
-}
-
-/// Ensures the existence of configuration files required by the application.
-pub fn ensure_config_files_exist(
-    base16_config_path: &Path,
-    base16_shell_theme_name_path: &Path,
-) -> Result<()> {
-    ensure_directory_exists(base16_config_path).with_context(|| {
-        format!(
-            "Failed to create config directory at {:?}",
-            base16_config_path
-        )
-    })?;
-    ensure_file_exists(base16_shell_theme_name_path).with_context(|| {
-        format!(
-            "Failed to create config file at {:?}",
-            base16_shell_theme_name_path
-        )
-    })?;
-
-    Ok(())
-}
-
-/// Reads the contents of a file and returns it as a string.
-fn read_file_to_string(path: &Path) -> Result<String> {
-    let mut file = File::open(path)?;
-    let mut contents = String::new();
-
-    file.read_to_string(&mut contents)?;
-
-    Ok(contents)
-}
-
-fn write_to_file(path: &Path, contents: &str) -> Result<()> {
-    if path.exists() {
-        fs::remove_file(&path)?;
-    }
-
-    let mut file = File::create(&path)?;
-    file.write_all(contents.as_bytes())?;
-
-    Ok(())
-}
 
 fn set_colorscheme_provided(
     theme_name: &String,
