+++ src/utils.rs
@@ -1,0 +1,69 @@
+use anyhow::{Context, Result};
+use std::fs::{self, File};
+use std::io::{Read, Write};
+use std::path::Path;
+
+/// Ensures that a directory exists, creating it if it does not.
+fn ensure_directory_exists<P: AsRef<Path>>(dir_path: P) -> Result<()> {
+    let path = dir_path.as_ref();
+
+    if !path.exists() {
+        fs::create_dir_all(path)
+            .with_context(|| format!("Failed to create directory at {:?}", path))?;
+    }
+
+    Ok(())
+}
+
+/// Ensures that a file exists, creating it if it does not.
+fn ensure_file_exists<P: AsRef<Path>>(file_path: P) -> Result<()> {
+    let path = file_path.as_ref();
+
+    if !path.exists() {
+        File::create(path).with_context(|| format!("Failed to create file at {:?}", path))?;
+    };
+
+    Ok(())
+}
+
+/// Ensures the existence of configuration files required by the application.
+pub fn ensure_config_files_exist(
+    base16_config_path: &Path,
+    base16_shell_theme_name_path: &Path,
+) -> Result<()> {
+    ensure_directory_exists(base16_config_path).with_context(|| {
+        format!(
+            "Failed to create config directory at {:?}",
+            base16_config_path
+        )
+    })?;
+    ensure_file_exists(base16_shell_theme_name_path).with_context(|| {
+        format!(
+            "Failed to create config file at {:?}",
+            base16_shell_theme_name_path
+        )
+    })?;
+
+    Ok(())
+}
+
+/// Reads the contents of a file and returns it as a string.
+pub fn read_file_to_string(path: &Path) -> Result<String> {
+    let mut file = File::open(path)?;
+    let mut contents = String::new();
+
+    file.read_to_string(&mut contents)?;
+
+    Ok(contents)
+}
+
+pub fn write_to_file(path: &Path, contents: &str) -> Result<()> {
+    if path.exists() {
+        fs::remove_file(&path)?;
+    }
+
+    let mut file = File::create(&path)?;
+    file.write_all(contents.as_bytes())?;
+
+    Ok(())
+}
